sahur_claude/sahur/sahur-batch/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sahur-batch"
version = "0.1.0"
description = "Dynamic issue processing instance for SAHUR"
readme = "README.md"
authors = [
    {name = "SAHUR Team"}
]
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.8.0",
    "websockets>=11.0.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.5",
    "pgvector>=0.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[tool.setuptools]
packages = ["sahur_batch"]

[project.scripts]
sahur-batch = "sahur_batch.main:main"
